Renamed api_service to cex_service
